% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIhSSALT.R
\name{CIhSSALT}
\alias{CIhSSALT}
\title{Confidence Intervals for hSSALT}
\usage{
CIhSSALT(data, n, MLEhSSALT_Obj, censoring = 1, tau, r = NULL, 
         monitoring = "continuous", delta = NULL, CImethod = "asymptotic", 
         alpha = 0.05, B = 1000, maxit = 1000, tol = 1e-8, language = "CPP", 
         parallel = FALSE, ncores = 2)
}
\arguments{
\item{data}{sample, a vector. The given data should be a censored vector with observations less than or equal to \code{n}. When censoring type is \code{2}, the length of \code{data} should be \code{r}.}

\item{n}{sample size, a positive integer.}

\item{MLEhSSALT_Obj}{an \code{MLEhSSALT} object, returned by \code{MLEhSSALT()}.}

\item{censoring}{\code{1} for Type-I censoring or \code{2} for Type-II censoring. Default value is \code{1}.}

\item{tau}{If censoring type is \code{1}, \code{tau} is a vector with length 2; if censoring type is \code{2}, \code{tau} is a positive numeric value.}

\item{r}{If censoring type is \code{2}, \code{r} provides the pre-specified number of failures, a positive integer.}

\item{monitoring}{\code{"continuous"} or \code{"interval"}. Default value is \code{"continuous"}. For interval monitoring, only equally spaced inspection is supported.}

\item{delta}{if interval monitoring, interval length, a positive numeric value. Default value is \code{NULL}.}

\item{CImethod}{\code{"asymptotic"}, \code{"percentile"} or \code{"bca"} for asymptotic CIs, bootstrap percentile CIs and bootstrap bias-corrected and accelerated (BCa) bootstrap intervals. Default value is \code{"asymptotic"}.}

\item{alpha}{significance level. Default value is \code{0.05}.}

\item{B}{number of bootstrap repetitions, a positive integer, default value is \code{1000}.}

\item{maxit}{The maximum number of iterations allowed, a positive integer. Only for bootstrap methods. Default value is \code{1000}.}

\item{tol}{Tolerance limit for declaring algorithm convergence based on the change between two consecutive iterations. Only for bootstrap methods. Default value is \code{1e-8}.}

\item{language}{\code{"R"} or \code{"CPP"}. Only for bootstrap methods. Default value is \code{"CPP"}.}

\item{parallel}{support parallel computation, a logical value. Only for bootstrap methods. Default value is \code{FALSE}.}

\item{ncores}{the number of cores that are used in parallelization, a positive integer.}
}
\value{
A \code{CIhSSALT} object that includes the type of returned CIs and the CIs for four parameters at a given significance level.
}
\description{
Provide interval estimation of a simple hSSALT model with exponential (continuous) or geometric (interval) distribution.
}
\examples{
MLE <- MLEhSSALT(data = hSSALTdata$data, n = 35, censoring = 1, tau = c(8, 20),
         theta21 = 1, theta22 = 8, p = 0.4)
ci <- CIhSSALT(data = hSSALTdata$data, n = 35, MLEhSSALT = MLE, tau = c(8, 20))

}
