% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomohSSALT.R
\name{HomohSSALT}
\alias{HomohSSALT}
\title{Homogeneity test for hSSALT}
\usage{
HomohSSALT(data, n, censoring = 1, tau, r = NULL, alpha = 0.05, M = 10000)
}
\arguments{
\item{data}{sample, a vector. The given data should be a censored vector with observations less than or equal to \code{n}. When censoring type is \code{2}, the length of \code{data} should be \code{r}.}

\item{n}{sample size, a positive integer.}

\item{censoring}{\code{1} for Type-I censoring or \code{2} for Type-II censoring. Default value is \code{1}.}

\item{tau}{If censoring type is \code{1}, \code{tau} is a vector with length 2; if censoring type is \code{2}, \code{tau} is a positive numeric value.}

\item{r}{If censoring type is \code{2}, \code{r} provides the pre-specified number of failures, a positive integer.}

\item{alpha}{significance level. Default value is \code{0.05}.}

\item{M}{number of simulations used to generate critical values, a positive integer. Default value is \code{10000}.}
}
\value{
An \code{hSSALTtest} object containing a hypothesis test table that reports the test statistic, the simulated critical value at the given significance level, the alternative hypothesis, and the test decision.
}
\description{
Perform a homogeneity test under the second stress level \code{s2} of a simple hSSALT model with exponential (continuous) distribution.
}
\examples{
sample <- rhSSALT(n = 30, tau = c(5, 10), theta1 = 10, theta21 = 5, theta22 = 8, p = 0.4)
test <- HomohSSALT(data = sample$`censored sample`, n = 30, tau = c(5, 10))

}
