% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEhSSALT.R
\name{MLEhSSALT}
\alias{MLEhSSALT}
\title{Maximum Likelihood Estimation for hSSALT}
\usage{
MLEhSSALT(data, n, censoring = 1, tau, r = NULL, monitoring = "continuous", 
          delta = NULL, theta21, theta22, p, maxit = 1000, 
          tol = 1e-8, language = "CPP")
}
\arguments{
\item{data}{sample, a vector. The given data should be a censored vector with observations less than or equal to \code{n}. When censoring type is \code{2}, the length of \code{data} should be \code{r}.}

\item{n}{sample size, a positive integer.}

\item{censoring}{\code{1} for Type-I censoring or \code{2} for Type-II censoring. Default value is \code{1}.}

\item{tau}{If censoring type is \code{1}, \code{tau} is a vector with length 2; if censoring type is \code{2}, \code{tau} is a positive numeric value.}

\item{r}{If censoring type is \code{2}, \code{r} provides the pre-specified number of failures, a positive integer.}

\item{monitoring}{\code{"continuous"} or \code{"interval"}. Default value is \code{"continuous"}. For interval monitoring, only equally spaced inspection is supported.}

\item{delta}{if interval monitoring, interval length, a positive numeric value. Default value is \code{NULL}.}

\item{theta21}{initial value of \code{theta21} for the EM algorithm, can be both a numeric value or a vector of values. For an initial-value vector, the (ultimate) value with the largest log-likelihood is returned as the MLE.}

\item{theta22}{initial value of \code{theta22} for the EM algorithm, can be both a numeric value or a vector of values.}

\item{p}{initial value of mixture proportion \code{p}, can be both a numeric value or a vector of values.}

\item{maxit}{The maximum number of iterations allowed, an integer. Default value is \code{1000}.}

\item{tol}{Tolerance limit for declaring algorithm convergence based on the change between two consecutive iterations. Default value is \code{1e-8}.}

\item{language}{\code{"R"} or \code{"CPP"}. Only for bootstrap methods. Default value is \code{"CPP"}.}
}
\description{
Provide point estimation of a simple hSSALT model with exponential (continuous) or geometric (interval) distribution.
}
\examples{
mle <- MLEhSSALT(data = hSSALTdata$data, n = 35, censoring = 1, tau = c(8, 20),
       theta21 = 1, theta22 = 8, p = 0.4)

}
\seealso{
\code{\link{CIhSSALT}}
}
